chore: refactor into multiple modules so it's easier to find things
