+++ packages/mongodb-dialects/mongosh/src/main/kotlin/com/mongodb/jbplugin/dialects/mongosh/MongoshDialectFormatter.kt
@@ -2,23 +2,16 @@ package com.mongodb.jbplugin.dialects.mongosh
 
 import com.mongodb.jbplugin.dialects.DialectFormatter
 import com.mongodb.jbplugin.dialects.OutputQuery
+import com.mongodb.jbplugin.dialects.mongosh.aggr.canEmitAggregate
+import com.mongodb.jbplugin.dialects.mongosh.aggr.emitAggregateBody
+import com.mongodb.jbplugin.dialects.mongosh.aggr.isAggregate
 import com.mongodb.jbplugin.dialects.mongosh.backend.MongoshBackend
+import com.mongodb.jbplugin.dialects.mongosh.query.emitCollectionReference
+import com.mongodb.jbplugin.dialects.mongosh.query.emitQueryFilter
+import com.mongodb.jbplugin.dialects.mongosh.query.emitSort
+import com.mongodb.jbplugin.dialects.mongosh.query.returnsACursor
 import com.mongodb.jbplugin.mql.*
 import com.mongodb.jbplugin.mql.components.*
-import com.mongodb.jbplugin.mql.components.HasFieldReference.FromSchema
-import com.mongodb.jbplugin.mql.components.HasFieldReference.Inferred
-import com.mongodb.jbplugin.mql.components.HasFieldReference.Unknown
-import com.mongodb.jbplugin.mql.parser.anyError
-import com.mongodb.jbplugin.mql.parser.components.aggregationStages
-import com.mongodb.jbplugin.mql.parser.components.allFiltersRecursively
-import com.mongodb.jbplugin.mql.parser.components.hasName
-import com.mongodb.jbplugin.mql.parser.components.whenIsCommand
-import com.mongodb.jbplugin.mql.parser.count
-import com.mongodb.jbplugin.mql.parser.filter
-import com.mongodb.jbplugin.mql.parser.map
-import com.mongodb.jbplugin.mql.parser.matches
-import com.mongodb.jbplugin.mql.parser.nth
-import com.mongodb.jbplugin.mql.parser.parse
 import io.github.z4kn4fein.semver.Version
 import org.owasp.encoder.Encode
 
@@ -27,8 +20,8 @@ object MongoshDialectFormatter : DialectFormatter {
         query: Node<S>,
         queryContext: QueryContext,
     ): OutputQuery {
-        val isAggregate = isAggregate(query)
-        val canEmitAggregate = canEmitAggregate(query)
+        val isAggregate = query.isAggregate()
+        val canEmitAggregate = query.canEmitAggregate()
 
         val outputString = MongoshBackend(
             prettyPrint =
@@ -60,14 +53,14 @@ object MongoshDialectFormatter : DialectFormatter {
                 emitFunctionName(query.component<IsCommand>()?.type?.canonical ?: "find")
             }
             emitFunctionCall(long = true, {
-                if (isAggregate(query)) {
+                if (query.isAggregate()) {
                     emitAggregateBody(query)
                 } else {
-                    emitQueryBody(query, firstCall = true)
+                    emitQueryFilter(query, firstCall = true)
                 }
             })
 
-            if (returnsACursor(query)) {
+            if (query.returnsACursor()) {
                 emitSort(query)
             }
         }.computeOutput()
@@ -126,296 +119,4 @@ object MongoshDialectFormatter : DialectFormatter {
     }
 
     override fun formatType(type: BsonType) = ""
-}
-
-private fun <S> MongoshBackend.emitQueryBody(
-    node: Node<S>,
-    firstCall: Boolean = false
-): MongoshBackend {
-    val named = node.component<Named>()
-    val fieldRef = node.component<HasFieldReference<S>>()
-    val valueRef = node.component<HasValueReference<S>>()
-    val hasFilter = node.component<HasFilter<S>>()
-    val isLong = allFiltersRecursively<S>().parse(node).orElse { emptyList() }.size > 3
-
-    if (hasFilter != null && fieldRef == null && valueRef == null && named == null) {
-        // 1. has children, nothing else (root node)
-        if (firstCall) {
-            emitObjectStart(long = isLong)
-        }
-
-        hasFilter.children.forEach {
-            emitQueryBody(it)
-            emitObjectValueEnd()
-        }
-        if (firstCall) {
-            emitObjectEnd(long = isLong)
-        }
-    } else if (hasFilter == null && fieldRef != null && valueRef != null && named == null) {
-        // 2. no children, only a field: value case
-        if (firstCall) {
-            emitObjectStart(long = isLong)
-        }
-        emitObjectKey(resolveFieldReference(fieldRef))
-        emitContextValue(resolveValueReference(valueRef, fieldRef))
-        if (firstCall) {
-            emitObjectEnd(long = isLong)
-        }
-    } else {
-        named?.let {
-// 3. children and named
-            if (named.name == Name.EQ) {
-// normal a: b case
-                if (firstCall) {
-                    emitObjectStart(long = isLong)
-                }
-                if (fieldRef != null) {
-                    emitObjectKey(resolveFieldReference(fieldRef))
-                }
-
-                if (valueRef != null) {
-                    emitContextValue(resolveValueReference(valueRef, fieldRef))
-                }
-
-                hasFilter?.children?.forEach {
-                    emitQueryBody(it)
-                    emitObjectValueEnd()
-                }
-
-                if (firstCall) {
-                    emitObjectEnd(long = isLong)
-                }
-            } else if (setOf( // 1st basic attempt, to improve in INTELLIJ-76
-                    Name.GT,
-                    Name.GTE,
-                    Name.LT,
-                    Name.LTE
-                ).contains(named.name) &&
-                valueRef != null
-            ) {
-// a: { $gt: 1 }
-                if (firstCall) {
-                    emitObjectStart(long = isLong)
-                }
-
-                if (fieldRef != null) {
-                    emitObjectKey(resolveFieldReference(fieldRef))
-                }
-
-                emitObjectStart()
-                emitObjectKey(registerConstant('$' + named.name.canonical))
-                emitContextValue(resolveValueReference(valueRef, fieldRef))
-                emitObjectEnd()
-
-                if (firstCall) {
-                    emitObjectEnd(long = isLong)
-                }
-            } else if (setOf(
-                    // 1st basic attempt, to improve in INTELLIJ-77
-                    Name.AND,
-                    Name.OR,
-                    Name.NOR,
-                ).contains(named.name)
-            ) {
-                if (firstCall) {
-                    emitObjectStart()
-                }
-                emitObjectKey(registerConstant('$' + named.name.canonical))
-                emitArrayStart(long = true)
-                hasFilter?.children?.forEach {
-                    emitObjectStart()
-                    emitQueryBody(it)
-                    emitObjectEnd()
-                    emitObjectValueEnd()
-                    if (prettyPrint) {
-                        emitNewLine()
-                    }
-                }
-                emitArrayEnd(long = true)
-                if (firstCall) {
-                    emitObjectEnd()
-                }
-            } else if (named.name == Name.NOT && hasFilter?.children?.size == 1) {
-                // the not operator is a special case
-                // because we receive it as:
-                // $not: { $field$: $condition$ }
-                // and it needs to be:
-                // $field$: { $not: $condition$ }
-                // we will do a JIT translation
-
-                var innerChild = hasFilter.children.first()
-                val operation = innerChild.component<Named>()
-                val fieldRef = innerChild.component<HasFieldReference<S>>()
-                val valueRef = innerChild.component<HasValueReference<S>>()
-
-                if (fieldRef == null) { // we are in an "and" / "or"...
-                    // so we use $nor instead
-                    emitQueryBody(
-                        Node(
-                            node.source,
-                            node.components<Component>().filterNot { it is Named } + Named(Name.NOR)
-                        )
-                    )
-                    return@let
-                }
-
-                if (operation == null && valueRef == null) {
-                    return@let
-                }
-
-                if (firstCall) {
-                    emitObjectStart()
-                }
-
-                // emit field name first
-                emitObjectKey(resolveFieldReference(fieldRef))
-                // emit the $not
-                emitObjectStart()
-                emitObjectKey(registerConstant('$' + "not"))
-                emitQueryBody(
-                    Node(
-                        innerChild.source,
-                        listOfNotNull(
-                            operation,
-                            valueRef
-                        )
-                    )
-                )
-                emitObjectEnd()
-
-                if (firstCall) {
-                    emitObjectEnd()
-                }
-            } else if (named.name != Name.UNKNOWN && fieldRef != null && valueRef != null) {
-                if (firstCall) {
-                    emitObjectStart(long = isLong)
-                }
-                emitObjectKey(resolveFieldReference(fieldRef))
-                emitObjectStart(long = isLong)
-                emitObjectKey(registerConstant('$' + named.name.canonical))
-                emitContextValue(resolveValueReference(valueRef, fieldRef))
-                emitObjectEnd(long = isLong)
-                if (firstCall) {
-                    emitObjectEnd(long = isLong)
-                }
-            }
-        }
-    }
-
-    return this
-}
-
-private fun <S> MongoshBackend.emitAggregateBody(node: Node<S>): MongoshBackend {
-    // here we can assume that we only have 1 single stage that is a match
-    val matchStage = node.component<HasAggregation<S>>()!!.children[0]
-    val filter = matchStage.component<HasFilter<S>>()?.children?.getOrNull(0)
-    val longFilter = filter?.component<HasFilter<S>>()?.children?.size?.let { it > 3 } == true
-
-    emitArrayStart(long = true)
-    emitObjectStart()
-    emitObjectKey(registerConstant('$' + "match"))
-    if (filter != null) {
-        emitObjectStart(long = longFilter)
-        emitQueryBody(filter)
-        emitObjectEnd(long = longFilter)
-    } else {
-        emitComment("No filter provided.")
-    }
-    emitObjectEnd()
-    emitArrayEnd(long = true)
-
-    return this
-}
-
-private fun <S> isAggregate(node: Node<S>): Boolean {
-    return whenIsCommand<S>(IsCommand.CommandType.AGGREGATE)
-        .map { true }
-        .parse(node).orElse { false }
-}
-
-private fun <S> returnsACursor(node: Node<S>): Boolean {
-    return whenIsCommand<S>(IsCommand.CommandType.FIND_MANY)
-        .map { true }
-        .parse(node).orElse { false }
-}
-
-private fun <S> canEmitAggregate(node: Node<S>): Boolean {
-    return aggregationStages<S>()
-        .matches(count<Node<S>>().filter { it >= 1 }.matches().anyError())
-        .nth(0)
-        .matches(hasName(Name.MATCH))
-        .map { true }
-        .parse(node).orElse { false }
-}
-
-private fun <S> MongoshBackend.resolveValueReference(
-    valueRef: HasValueReference<S>,
-    fieldRef: HasFieldReference<S>?
-) = when (val ref = valueRef.reference) {
-    is HasValueReference.Constant -> registerConstant(ref.value)
-    is HasValueReference.Inferred -> registerConstant(ref.value)
-    is HasValueReference.Runtime -> registerVariable(
-        (fieldRef?.reference as? FromSchema)?.fieldName ?: "value",
-        ref.type
-    )
-
-    else -> registerVariable(
-        "queryField",
-        BsonAny
-    )
-}
-
-private fun <S> MongoshBackend.resolveFieldReference(fieldRef: HasFieldReference<S>) =
-    when (val ref = fieldRef.reference) {
-        is FromSchema -> registerConstant(ref.fieldName)
-        is Inferred<S> -> registerConstant(ref.fieldName)
-        is HasFieldReference.Computed<S> -> registerConstant(ref.fieldName)
-        is Unknown -> registerVariable("field", BsonAny)
-    }
-
-private fun <S> MongoshBackend.emitCollectionReference(collRef: HasCollectionReference<S>?): MongoshBackend {
-    when (val ref = collRef?.reference) {
-        is HasCollectionReference.OnlyCollection -> {
-            emitDatabaseAccess(registerVariable("database", BsonString))
-            emitCollectionAccess(registerConstant(ref.collection))
-        }
-
-        is HasCollectionReference.Known -> {
-            emitDatabaseAccess(registerConstant(ref.namespace.database))
-            emitCollectionAccess(registerConstant(ref.namespace.collection))
-        }
-
-        else -> {
-            emitDatabaseAccess(registerVariable("database", BsonString))
-            emitCollectionAccess(registerVariable("collection", BsonString))
-        }
-    }
-
-    return this
-}
-
-private fun <S> MongoshBackend.emitSort(query: Node<S>): MongoshBackend {
-    val sortComponent = query.component<HasSorts<S>>()
-    if (sortComponent == null) {
-        return this
-    }
-
-    fun generateSortKeyVal(node: Node<S>): MongoshBackend {
-        val fieldRef = node.component<HasFieldReference<S>>() ?: return this
-        val valueRef = node.component<HasValueReference<S>>() ?: return this
-
-        emitObjectKey(resolveFieldReference(fieldRef))
-        emitContextValue(resolveValueReference(valueRef, fieldRef))
-        return emitObjectValueEnd()
-    }
-
-    emitPropertyAccess()
-    emitFunctionName("sort")
-    return emitFunctionCall(long = false, {
-        emitObjectStart(long = false)
-        for (sortCriteria in sortComponent.children) {
-            generateSortKeyVal(sortCriteria)
-        }
-        emitObjectEnd(long = false)
-    })
 }
